Split up Flow.js into distinct classes and rely upon build-time assembly.
- `$` shorthand symbol was removed in favor of `this` and context has been taken care of (including for event handlers)
- _real_ class methods replace the former simili-constructor syntax
- started replacing a couple `each()` by `for..of` when it helped getting rid of significant context switch complexity.
- FlowFile and evalOpts (counter-intuitive) bindings into Flow namespace had to be dirtily reproduced because the Grunt/Karma/Jasmine hardly support full ES6 import within spec files.
- tests adapted
- flow.support is removed
  1. Because if there is no `window`, user are on their own and there is no reason keeping node.js from loading the code
  2. Because `support` rather depend on the transpilation and polyfills part of the final build issues by the toolchain
  => Avoid cluttering bootstrap with a runtime check.
