+++ test/evalOptsSpec.js
@@ -1,17 +1,19 @@
+// import evalOpts from '/test/../src/tools';
+
 describe('evalOpts', function () {
 
   it('should return same object for non functions', function() {
     var obj = {};
-    expect(Flow.evalOpts(obj)).toBe(obj);
+    expect(evalOpts(obj)).toBe(obj);
   });
   it('should return same type for non functions', function() {
-    expect(Flow.evalOpts(5)).toBe(5);
+    expect(evalOpts(5)).toBe(5);
   });
   it('should evaluate function', function() {
-    expect(Flow.evalOpts(function () {return 5;})).toBe(5);
+    expect(evalOpts(function () {return 5;})).toBe(5);
   });
   it('should evaluate function with given arguments', function() {
     var obj = {};
-    expect(Flow.evalOpts(function (a) {return a;}, obj)).toBe(obj);
+    expect(evalOpts(function (a) {return a;}, obj)).toBe(obj);
   });
 });
